TagReader: Fix reading nested/compound lists
StandardTagStreamHandler: Fix endianness of UTF length being flipped (tried to read as LE in BE mode, tried to read as BE in LE mode)
Add read test

+++ src/main/java/io/github/speedbridgemc/nibblet/TagIO.java
@@ -168,14 +168,14 @@ public final class TagIO {
             reader.beginIntArray();
             int[] iArr = new int[reader.listSize()];
             for (int i = 0; reader.listHasNext(); i++)
-                iArr[i] = reader.nextByte();
+                iArr[i] = reader.nextInt();
             reader.endIntArray();
             return IntArrayTag.copyOf(iArr);
         case LONG_ARRAY:
             reader.beginLongArray();
             long[] lArr = new long[reader.listSize()];
             for (int i = 0; reader.listHasNext(); i++)
-                lArr[i] = reader.nextByte();
+                lArr[i] = reader.nextLong();
             reader.endLongArray();
             return LongArrayTag.copyOf(lArr);
         default:
